+++ crates/astrid-storage/src/volume/hosted/tests.rs
@@ -1,4 +1,4 @@
-use super::reclaim::{ReclaimStage, set_stage_hook, set_test_unlock_hook};
+use super::reclaim::set_test_unlock_hook;
 use super::*;
 use crate::volume::VolumeFile;
 
@@ -6,7 +6,7 @@ fn append_valid_uncommitted_write(path: &std::path::Path, sequence: u64, payload
     append_valid_record(path, sequence, Operation::Write, b"objects", 0, payload);
 }
 
-fn append_valid_record(
+pub(super) fn append_valid_record(
     path: &std::path::Path,
     sequence: u64,
     operation: Operation,
@@ -224,356 +224,6 @@ fn legacy_empty_commit_recovers_without_hashing_write_payloads() {
         .read_region_at(&VolumeRegion::new("objects").unwrap(), 0, &mut recovered)
         .unwrap();
     assert_eq!(&recovered, b"legacy");
-}
-
-fn seeded_volume(path: &std::path::Path) -> Arc<HostedFileVolume> {
-    let volume = HostedFileVolume::open(path).unwrap();
-    let region = VolumeRegion::new("objects").unwrap();
-    volume.create_region(&region, true).unwrap();
-    volume
-        .write_region_at(&region, 0, &vec![0xAA; 256 * 1024])
-        .unwrap();
-    volume.write_region_at(&region, 0, b"small").unwrap();
-    volume.set_region_len(&region, 5).unwrap();
-    volume.sync().unwrap();
-    volume
-}
-
-fn snapshot_reclaim_at(stage: ReclaimStage) -> (tempfile::TempDir, std::path::PathBuf) {
-    let temporary = tempfile::tempdir().unwrap();
-    let path = temporary.path().join("astrid.volume");
-    let snapshot = temporary.path().join(format!(
-        "snapshot-{}",
-        match stage {
-            ReclaimStage::BoundImagePublished => "bound-root",
-            ReclaimStage::ReplacementDurable => "unflipped",
-            ReclaimStage::RootPublished => "replacement-root",
-            ReclaimStage::FinalImageDurable => "final-image",
-            ReclaimStage::FinalRootPublished => "final-root",
-            ReclaimStage::Truncated => "truncated",
-        }
-    ));
-    let volume = seeded_volume(&path);
-    let source = path.clone();
-    let target = snapshot.clone();
-    set_stage_hook(
-        path.clone(),
-        Box::new(move |seen| {
-            if seen == stage {
-                std::fs::copy(&source, &target)?;
-            }
-            Ok(())
-        }),
-    );
-    let reclaiming = Arc::clone(&volume);
-    let result = std::thread::spawn(move || reclaiming.reclaim_same_inode())
-        .join()
-        .unwrap();
-    result.unwrap();
-    assert!(snapshot.exists(), "reclaim did not reach {stage:?}");
-    (temporary, snapshot)
-}
-
-fn read_objects(volume: &HostedFileVolume) -> [u8; 5] {
-    let region = VolumeRegion::new("objects").unwrap();
-    let mut bytes = [0_u8; 5];
-    volume.read_region_at(&region, 0, &mut bytes).unwrap();
-    bytes
-}
-
-#[test]
-fn inode_stable_reclaim_shrinks_the_selected_generation() {
-    let temporary = tempfile::tempdir().unwrap();
-    let path = temporary.path().join("astrid.volume");
-    let volume = seeded_volume(&path);
-    let before = std::fs::metadata(&path).unwrap().len();
-    volume.reclaim_same_inode().unwrap();
-    let after = std::fs::metadata(&path).unwrap().len();
-    assert!(after < before, "{before} did not shrink to {after}");
-    assert_eq!(read_objects(&volume), *b"small");
-    drop(volume);
-    let reopened = HostedFileVolume::open(&path).unwrap();
-    assert_eq!(read_objects(&reopened), *b"small");
-}
-
-#[test]
-fn same_inode_reclaim_preserves_an_empty_volume() {
-    let temporary = tempfile::tempdir().unwrap();
-    let path = temporary.path().join("astrid.volume");
-    let volume = HostedFileVolume::open(&path).unwrap();
-    assert_eq!(std::fs::metadata(&path).unwrap().len(), 8);
-    volume.reclaim_same_inode().unwrap();
-    assert_eq!(std::fs::metadata(&path).unwrap().len(), 8);
-    assert_eq!(volume.list_regions("").unwrap(), Vec::new());
-    drop(volume);
-    let reopened = HostedFileVolume::open(&path).unwrap();
-    assert_eq!(reopened.list_regions("").unwrap(), Vec::new());
-    assert_eq!(std::fs::metadata(&path).unwrap().len(), 8);
-}
-
-#[test]
-fn generation_root_recovers_after_uncommitted_tail() {
-    let temporary = tempfile::tempdir().unwrap();
-    let path = temporary.path().join("astrid.volume");
-    let region = VolumeRegion::new("objects").unwrap();
-    let volume = seeded_volume(&path);
-    volume.reclaim_same_inode().unwrap();
-    let mut file = OpenOptions::new().read(true).open(&path).unwrap();
-    let selected = recover::recover_container(&mut file).unwrap();
-    drop(file);
-    assert!(selected.generation > 0);
-    let authority_len = selected.authority_len;
-
-    volume
-        .detach_after_uncommitted_write_for_test(&region, b"pending")
-        .unwrap();
-    drop(volume);
-    assert!(std::fs::metadata(&path).unwrap().len() > authority_len);
-
-    let reopened = HostedFileVolume::open(&path).unwrap();
-    assert_eq!(read_objects(&reopened), *b"small");
-    assert_eq!(std::fs::metadata(&path).unwrap().len(), authority_len);
-}
-
-#[test]
-fn same_generation_stale_pointer_does_not_block_new_root() {
-    let temporary = tempfile::tempdir().unwrap();
-    let path = temporary.path().join("astrid.volume");
-    let region = VolumeRegion::new("objects").unwrap();
-    let volume = seeded_volume(&path);
-    volume.reclaim_same_inode().unwrap();
-    let mut file = OpenOptions::new().read(true).open(&path).unwrap();
-    let selected = recover::recover_container(&mut file).unwrap();
-    let stale_footer = selected.durable_len;
-    let stale_root_base = selected.root_base;
-    drop(file);
-
-    volume.write_region_at(&region, 0, b"new").unwrap();
-    volume.sync().unwrap();
-    let mut file = OpenOptions::new()
-        .read(true)
-        .write(true)
-        .open(&path)
-        .unwrap();
-    let selected = recover::recover_container(&mut file).unwrap();
-    let generation = selected.generation;
-    let root_base = selected.root_base;
-    assert!(selected.durable_len > stale_footer);
-
-    // Reproduce the crash point at which the first (slot-zero) copy names the
-    // new footer while its same-generation sibling still names the old one.
-    recover::write_root_pointer(&mut file, generation, stale_root_base, stale_footer, true)
-        .unwrap();
-    file.sync_all().unwrap();
-    drop(file);
-    volume.detach_for_test().unwrap();
-    drop(volume);
-
-    let reopened = HostedFileVolume::open(&path).unwrap();
-    let mut bytes = [0_u8; 3];
-    reopened.read_region_at(&region, 0, &mut bytes).unwrap();
-    assert_eq!(&bytes, b"new");
-    assert!(
-        !recover::recover_container(&mut OpenOptions::new().read(true).open(&path).unwrap())
-            .unwrap()
-            .pointer_slot
-    );
-    assert_eq!(root_base, stale_root_base);
-}
-
-#[test]
-fn unflipped_replacement_image_leaves_the_old_generation_current() {
-    let (temporary, snapshot) = snapshot_reclaim_at(ReclaimStage::ReplacementDurable);
-    let physical_len = std::fs::metadata(&snapshot).unwrap().len();
-    let mut file = OpenOptions::new()
-        .read(true)
-        .write(true)
-        .open(&snapshot)
-        .unwrap();
-    let recovery = recover::recover_container(&mut file).unwrap();
-    // Generation one is the durable copy of the pre-reclaim image. The
-    // replacement's footer is beyond that selected authority, so an EOF scan
-    // cannot silently promote it before its own root flip.
-    assert_eq!(recovery.generation, 1);
-    assert!(recovery.root_base < recovery.last_commit_offset);
-    assert!(physical_len > recovery.authority_len);
-    let volume = HostedFileVolume::open(&snapshot).unwrap();
-    assert_eq!(read_objects(&volume), *b"small");
-    assert_eq!(
-        std::fs::metadata(&snapshot).unwrap().len(),
-        recovery.authority_len
-    );
-    drop(volume);
-    assert_eq!(temporary.path().read_dir().unwrap().count(), 2);
-}
-
-#[test]
-fn replacement_root_is_selected_before_the_final_image_exists() {
-    let (temporary, snapshot) = snapshot_reclaim_at(ReclaimStage::RootPublished);
-    let volume = HostedFileVolume::open(snapshot).unwrap();
-    assert_eq!(read_objects(&volume), *b"small");
-    drop(volume);
-    assert_eq!(temporary.path().read_dir().unwrap().count(), 2);
-}
-
-#[test]
-fn durable_final_image_is_ignored_until_its_root_is_selected() {
-    let (temporary, snapshot) = snapshot_reclaim_at(ReclaimStage::FinalImageDurable);
-    let volume = HostedFileVolume::open(snapshot).unwrap();
-    assert_eq!(read_objects(&volume), *b"small");
-    drop(volume);
-    assert_eq!(temporary.path().read_dir().unwrap().count(), 2);
-}
-
-#[test]
-fn selected_final_root_survives_until_truncate_completes() {
-    let (temporary, flipped) = snapshot_reclaim_at(ReclaimStage::FinalRootPublished);
-    let untruncated = std::fs::metadata(&flipped).unwrap().len();
-    let volume = HostedFileVolume::open(flipped).unwrap();
-    assert_eq!(read_objects(&volume), *b"small");
-    drop(volume);
-    let (complete_temporary, complete) = snapshot_reclaim_at(ReclaimStage::Truncated);
-    let truncated = std::fs::metadata(&complete).unwrap().len();
-    assert!(truncated < untruncated, "{untruncated} -> {truncated}");
-    assert_eq!(temporary.path().read_dir().unwrap().count(), 2);
-    assert_eq!(complete_temporary.path().read_dir().unwrap().count(), 2);
-}
-
-#[test]
-fn torn_root_candidates_are_not_admitted_as_authority() {
-    let (_temporary, snapshot) = snapshot_reclaim_at(ReclaimStage::Truncated);
-    let mut bytes = std::fs::read(&snapshot).unwrap();
-    // Damage the checksum in each fixed root copy. Neither byte sequence can
-    // masquerade as generation zero, and the EOF footer alone is no longer a
-    // root authority once a generation pointer has existed.
-    bytes[VOLUME_MAGIC.len() + ROOT_SLOT_BYTES - 1] ^= 0x80;
-    bytes[ROOT_BYTES - 1] ^= 0x80;
-    std::fs::write(&snapshot, bytes).unwrap();
-    let error = HostedFileVolume::open(snapshot).unwrap_err();
-    assert_eq!(error.kind(), io::ErrorKind::InvalidData);
-    assert!(
-        error
-            .to_string()
-            .contains("invalid Astrid volume root pointer"),
-        "{error}"
-    );
-}
-
-#[test]
-fn higher_valid_root_generation_wins() {
-    let (_temporary, snapshot) = snapshot_reclaim_at(ReclaimStage::Truncated);
-    let mut file = OpenOptions::new()
-        .read(true)
-        .write(true)
-        .open(&snapshot)
-        .unwrap();
-    let footer_offset = file.metadata().unwrap().len() - recover::FOOTER_BYTES as u64;
-    recover::write_root_pointer(&mut file, 9, ROOT_BYTES as u64, footer_offset, false).unwrap();
-    recover::write_root_pointer(&mut file, 7, ROOT_BYTES as u64, footer_offset, true).unwrap();
-    file.sync_all().unwrap();
-    let volume = HostedFileVolume::open(snapshot).unwrap();
-    assert_eq!(read_objects(&volume), *b"small");
-}
-
-#[test]
-fn generation_zero_eof_footer_remains_recoverable_and_reclaimable() {
-    let temporary = tempfile::tempdir().unwrap();
-    let path = temporary.path().join("astrid.volume");
-    std::fs::write(&path, VOLUME_MAGIC).unwrap();
-    append_valid_record(&path, 1, Operation::Create, b"objects", 0, &[]);
-    append_valid_record(&path, 2, Operation::Write, b"objects", 0, b"small");
-    append_valid_record(
-        &path,
-        3,
-        Operation::Commit,
-        COMMIT_REGION.as_bytes(),
-        0,
-        &encode_region_snapshot_for_test(
-            VOLUME_MAGIC.len() as u64
-                + 2 * u64::try_from(RECORD_FIXED_BYTES + b"objects".len()).unwrap(),
-        ),
-    );
-    let durable_len = std::fs::metadata(&path).unwrap().len();
-    let snapshot = encode_region_snapshot_for_test(
-        VOLUME_MAGIC.len() as u64
-            + 2 * u64::try_from(RECORD_FIXED_BYTES + b"objects".len()).unwrap(),
-    );
-    let commit_length = RECORD_FIXED_BYTES + b"system/volume-commit".len() + snapshot.len();
-    let commit_offset = durable_len - u64::try_from(commit_length).unwrap();
-    {
-        let mut file = OpenOptions::new().write(true).open(&path).unwrap();
-        recover::write_footer(&mut file, commit_offset, durable_len, 3).unwrap();
-        file.sync_all().unwrap();
-    }
-    let volume = HostedFileVolume::open(&path).unwrap();
-    assert_eq!(read_objects(&volume), *b"small");
-    drop(volume);
-    let reopened = HostedFileVolume::open(&path).unwrap();
-    assert_eq!(read_objects(&reopened), *b"small");
-}
-
-#[test]
-fn receipt_region_survives_all_same_inode_reclaim_stages() {
-    let temporary = tempfile::tempdir().unwrap();
-    let path = temporary.path().join("astrid.volume");
-    let volume = seeded_volume(&path);
-    let receipt = VolumeRegion::new("system/gc-outbox/receipt.ready").unwrap();
-    volume.create_region(&receipt, true).unwrap();
-    volume.write_region_at(&receipt, 0, b"proof").unwrap();
-    volume.sync().unwrap();
-    volume.reclaim_same_inode().unwrap();
-    let mut proof = [0_u8; 5];
-    volume.read_region_at(&receipt, 0, &mut proof).unwrap();
-    assert_eq!(&proof, b"proof");
-    drop(volume);
-    let reopened = HostedFileVolume::open(&path).unwrap();
-    reopened.read_region_at(&receipt, 0, &mut proof).unwrap();
-    assert_eq!(&proof, b"proof");
-}
-
-#[test]
-fn physical_credit_waits_until_the_final_root_and_truncate() {
-    let temporary = tempfile::tempdir().unwrap();
-    let path = temporary.path().join("astrid.volume");
-    let volume = seeded_volume(&path);
-    let before = std::fs::metadata(&path).unwrap().len();
-    let source = path.clone();
-    set_stage_hook(
-        path.clone(),
-        Box::new(move |stage| {
-            let current = std::fs::metadata(&source).unwrap().len();
-            match stage {
-                ReclaimStage::ReplacementDurable
-                | ReclaimStage::BoundImagePublished
-                | ReclaimStage::RootPublished
-                | ReclaimStage::FinalImageDurable
-                | ReclaimStage::FinalRootPublished => assert!(current >= before),
-                ReclaimStage::Truncated => assert!(current < before),
-            }
-            Ok(())
-        }),
-    );
-    volume.reclaim_same_inode().unwrap();
-    let available = volume.available_space().unwrap().expect("host capacity");
-    assert!(available > 0);
-}
-
-fn encode_region_snapshot_for_test(physical_offset: u64) -> Vec<u8> {
-    let mut regions = BTreeMap::new();
-    regions.insert(
-        VolumeRegion::new("objects").unwrap(),
-        RegionState {
-            length: 5,
-            extents: BTreeMap::from([(
-                0,
-                Extent {
-                    logical_end: 5,
-                    physical_offset,
-                },
-            )]),
-        },
-    );
-    recover::encode_region_snapshot(&regions).unwrap()
 }
 
 #[test]
